requirements.txt
# Python dependencies required for development
blinker
click
Flask
Flask-Cors
Flask-Mail
itsdangerous
Jinja2
MarkupSafe
pip
python-dotenv
setuptools
Werkzeug

